requirements.txt
openai-agents>=0.2.0
python-dotenv>=1.0.1
pdfplumber>=0.11.4
xlrd>=2.0.1
openpyxl>=3.1.5

# Dev tools
black>=24.8.0
ruff>=0.6.8
pylint>=3.2.6
pyright>=1.1.377
pre-commit>=3.7.1

